Update coc to disambiguate Code to Code For Good
Since this is a Code of Conduct, and the org name also has the word Code, "Code" seems ambiguous here or at least might get in the way of a reading. I think I've interpreted it right in that it should be referencing Code for Good in these two modified locations.